Switch inputs.js to a node script
