+++ package.json
@@ -13,6 +13,7 @@
   },
   "scripts": {
     "test": "jest",
+    "test:watch": "jest --watch --notify",
     "lint": "eslint src .github --ignore-pattern '!.github'",
     "prettier": "prettier --write src"
   },
@@ -28,6 +29,7 @@
     "set-value": ">=4.0.1"
   },
   "dependencies": {
+    "@actions/core": "^1.9.1",
     "@actions/github": "^5.0.3",
     "@sentry/node": "^7.12.0"
   }
